Simplify the API
Simplify the API by removing `hasUnsetVariables`, `hasVariable`, and `setVariable` and replacing these methods with a single parameter to `evaluate`. `evaluate` now takes a single dictionary mapping variables to their values.

This change updates the tests to account for these changes and fixes the exception classes to correctly set the message member so that the tests can now all pass.

Since this is already a breaking change, browserify is opportunistically updated to the next major version.